Changing Voxel-Phase to use unsigned int from float and all other support code for importing the OIM data

+++ Source/AIM/Common/Voxel.h
@@ -78,7 +78,7 @@ class AIMCOMMON_EXPORT Voxel
     double euler1;
     double euler2;
     double euler3;
-    float phase;
+    unsigned int phase;
     int neighbor;
     int numowners;
     double grainmisorientation;
